# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LADO"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions": {
      "get": {
        "summary": "Sessions",
        "operationId": "sessions_api_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionInfo"
                  },
                  "type": "array",
                  "title": "Response Sessions Api Sessions Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "summary": "Events",
        "description": "The change feed as Server-Sent Events (lado.server.feed). The position is the\nLast-Event-ID header (the browser's own reconnect) or, without it, `after`.",
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "last-event-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{name}/agents": {
      "get": {
        "summary": "Agents",
        "operationId": "agents_api_sessions__name__agents_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentInfo"
                  },
                  "title": "Response Agents Api Sessions  Name  Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{name}/agents/{agent}/history": {
      "get": {
        "summary": "History",
        "description": "The agent's window: its last `lines` lines, for the UI's read-only history, and\nwhether the agent shows a full-screen program, whose history is inside it.",
        "operationId": "history_api_sessions__name__agents__agent__history_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "agent",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent"
            }
          },
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 1,
              "default": 2000,
              "title": "Lines"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "status": {
            "type": "string",
            "enum": [
              "starting",
              "busy",
              "idle",
              "waiting",
              "stopped"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "name",
          "role",
          "provider",
          "status"
        ],
        "title": "AgentInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "version"
        ],
        "title": "Health"
      },
      "History": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "alternate": {
            "type": "boolean",
            "title": "Alternate"
          }
        },
        "type": "object",
        "required": [
          "text",
          "alternate"
        ],
        "title": "History"
      },
      "SessionInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "repo": {
            "type": "string",
            "title": "Repo"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "agents": {
            "type": "integer",
            "title": "Agents"
          }
        },
        "type": "object",
        "required": [
          "name",
          "repo",
          "status",
          "agents"
        ],
        "title": "SessionInfo"
      },
      "SessionStatus": {
        "type": "string",
        "enum": [
          "stopped",
          "running",
          "tmux_gone",
          "loop_down"
        ],
        "title": "SessionStatus",
        "description": "Whether a session runs, as `lado ls` and the UI show it."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}